Update following Daniel's comments
- On the third paragraph, there is a sentence that doesn't make much
sense. I think that it was supposed to be the titles of the sections
that can be seen below. I highlight the text that should be removed:
While we are organizing only the mentoring lunch, we expect that some
of you might want to have a more sustained relationship with the
mentors and we invite you to ask the mentors at lunch about such a
possibility. Target groupsGraduate Students

- Could you please add the following at the end?
Preliminary list of mentors
A list with the mentors will appear soon.
Organizers
Daniel Garijo, Information Sciences Institute
Abraham Bernstein, University of Zurich